% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addStepBounds.R
\name{addStepBounds}
\alias{addStepBounds}
\title{Plot Partial Identification Bounds}
\usage{
addStepBounds(
  F1,
  F2,
  x.limit = NULL,
  col.bounds = "blue",
  alpha.bounds = 0.1,
  lty = 1,
  col.lines = NA
)
}
\arguments{
\item{F1, }{F2 Lower and upper partial identification bounds. F1 and F2 must be step functios.}

\item{x.limit}{A vector of length 2 specifying the minimum and maximum x-values for the plot.
If NULL, defaults to 20 times the range of the x-values in the object.}

\item{col.bounds}{Color of the shaded area representing the partial identification bounds.}

\item{alpha.bounds}{Transparency level of the shaded area.}
}
\description{
This function plots the partial identification bounds for a given range of x values
based on an object of class "ecdfPI". It creates a shaded area between the lower and upper bounds.
}
