% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdfdiffY1Y0.R
\name{cdfdiffY1Y0}
\alias{cdfdiffY1Y0}
\title{Calculate the Mass and Distribution of the Difference U = Y1 - Y0}
\usage{
cdfdiffY1Y0(Copula, FY1, FY0, suppY1 = NULL, suppY0 = NULL)
}
\arguments{
\item{Copula}{A copula function.}

\item{FY1}{Function representing the CDF of Y1.}

\item{FY0}{Function representing the CDF of Y0.}

\item{suppY1}{Support of Y1.}

\item{suppY0}{Support of Y0.}
}
\value{
A list containing two elements:
\item{jointdist}{A data frame with the joint distribution of (Y1, Y0) under the specified copula.}
\item{massU}{A data frame with the distribution of U = Y1 - Y0.}
\item{cdfU}{CDF of U = Y1 - Y0.}
}
\description{
This function computes the mass and cumulative distribution function (CDF)
of the difference U = Y1 - Y0 using a given copula.
}
\examples{
# Define copulas: 
W <- function(a,b) { pmax(a+b-1,0)}
M <- function(a,b) { pmin(a,b)}
PI <- function(a,b) { a*b }

Y1 <- iris$Sepal.Length[iris$Species == "setosa"]
Y0 <- iris$Sepal.Length[iris$Species == "virginica"]
cdfUPI <- cdfdiffY1Y0(Copula = PI,
                      FY1 = ecdf(Y1),
                      FY0 = ecdf(Y0),
                      suppY1 = sort(unique(Y1)),
                      suppY0 = sort(unique(Y0)) )

plot( cdfUPI$cdfU )


}
