% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTEprobdata.R
\name{CTEprobdata}
\alias{CTEprobdata}
\title{Calculate the Causal Treatment Effect (CTE) distribution for real data with dichotomous outcome}
\usage{
CTEprobdata(
  Yobs,
  Zobs,
  X = NULL,
  p = c(0:60)/60,
  y = NULL,
  covariates = F,
  alpha = 0.05
)
}
\arguments{
\item{Yobs}{A vector of observed outcomes. A dichotomous variable. If \code{Yobs} is not dichotomous, it is
transformed into a dichotomous variable using the threshold \code{y}.}

\item{Zobs}{A vector of observed treatments (\code{0} or \code{1}).}

\item{X}{A matrix of covariates for each individual. It can be used for calculating conditional \eqn{CTE(0|X=x)} distributions.}

\item{p}{A vector of prior propensity values for which the CTE distribution is calculated.}

\item{y}{If \code{NULL}, \code{Yobs} is treated as a dichotomous variable. If provided, it is used as a threshold to create
a dichotomous outcome variable: \code{Y > y}.}

\item{covariates}{A logical value indicating whether the \eqn{CTE(0)} distribution should be calculated for each subset defined by the covariates
(\code{TRUE}) or for the whole dataset (\code{FALSE}).}
}
\value{
If \code{covariates} is \code{FALSE}, \code{CTEprobdata} returns a list with the CTE(0) distribution over the possible worlds
for each value in the \code{p} vector. The list also includes \code{probMZ}, containing probabilities identified in the
\eqn{\mathcal{M} \times \{0, 1\}} space:
\itemize{
\item{\code{P(K=1|Z=1)}}: Probability that \eqn{K=1} given \eqn{Z=1}.
\item{\code{P(K=1|Z=0)}}: Probability that \eqn{K=1} given \eqn{Z=0}.
\item{\code{P(Y=1|Z=1,K=1)}}: Probability that \eqn{Y=1} given \eqn{Z=1} and \eqn{K=1}.
\item{\code{P(Y=1|Z=0,K=1)}}: Probability that \eqn{Y=1} given \eqn{Z=0} and \eqn{K=1}.
\item{\code{p0}}: The prior propensity where the probability of a positive causal effect is equal to the probability of a negative causal effect.
}

If \code{covariates} is \code{TRUE}, \code{CTEprobdata} returns a list for each subset defined by the covariates \code{X}.
The name of each list corresponds to the values of the covariates. If the covariate is a factor, the name of each list element
refers to the level number rather than the label of the level. For each partition, \code{probMZ} also includes the number of labels
for the partition (\code{M}) and the number of missing data (\code{NA}) in the outcome.
}
\description{
This function calculates the Causal Treatment Effect (CTE) distribution for real data with a dichotomous outcome variable.
}
\details{
\code{CTEprobdata} accepts observed data \code{(Yobs, Zobs, X)}. Each record corresponds to an observation for an individual.
If \eqn{\mathcal{M}} represents the set of labels for the individuals, then \code{(Yobs, Zobs, X)} are variables in \eqn{\mathcal{M}}.
The outcome \eqn{Y} is defined over the space \eqn{\mathcal{M} \times \{0, 1\}}. \eqn{Y} has at least \code{M} missing data points (counterfactuals).
Each missing outcome can take values \code{0} or \code{1}.

Consider the space of all possible worlds, where for each world there exists an outcome \eqn{Y^*}, where \eqn{Y^*} coincides
with \code{Yobs} for the observed statistical units. For each \eqn{Y^*}, the \eqn{CTE(0)} is calculated. \eqn{CTE(0)} is a random
variable over all possible worlds.

The value \eqn{p} represents the prior probability that each unobserved outcome is \code{1}. This prior probability is
called the "propensity for success." Assuming a common \eqn{p} for all unobserved outcomes implies the absence of a causal effect.
Based on \eqn{p}, the probability of each possible world is determined. If all possible worlds are equally likely, then \eqn{p = 0.5},
implying no causal effect. This assumption is unrealistic when the outcome has a probability of success significantly
different from \code{0.5}.

The \eqn{CTE(0)} distribution over the space of possible worlds is calculated for different values of the prior propensity \eqn{p}.
This allows for the assessment of the impact of the assumption about \eqn{p} on the CTE(0) distribution, and consequently,
the potential existence of a positive or negative causal effect.

The function \code{plotCTEprob} can be used to visualize the probabilities of positive, negative, or no causal effect
for a given prior propensity \eqn{p}.
}
