% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecdfPI.R
\name{ecdfPI}
\alias{ecdfPI}
\title{Calculate ECDF with Partial Identification Bounds (considering NAs)}
\usage{
ecdfPI(x, weight = NULL)
}
\arguments{
\item{x}{A numeric vector.}

\item{weight}{A numeric vector of weights. The sum of weights must be 1.}
}
\value{
An object of class \code{ecdfPI}, which is a list with:
\item{ecdf}{An ECDF function of class \code{ecdf}.}
\item{F1}{Lower bound of the ECDF with class \code{ecdfPIbound}.}
\item{F2}{Upper bound of the ECDF with class \code{ecdfPIbound}.}
}
\description{
This function calculates the empirical cumulative distribution function (ECDF)
along with its partial identification bounds under no assumptions.
It is a modification of the base \code{ecdf()} function and uses the class \code{ecdfPI}
to allow a specific plot method for these objects (\code{plot.ecdfPI}).
}
\examples{
x <- c(1, 2, 3, NA, 4, 5)
result <- ecdfPI(x)
plot(result)
plot(result$ecdf)
lines(result$F1, vertical=F, pch=19, col = "blue")
lines(result$F2, vertical=F, pch=19, col = "red")

}
