% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ex.R
\name{Ex}
\alias{Ex}
\title{Compute the Expectation for a Discrete CDF}
\usage{
Ex(sf)
}
\arguments{
\item{sf}{A stepfun object representing the discrete CDF.}
}
\value{
A list with the following components:
\describe{
\item{E}{The total expectation of the distribution.}
\item{Eplus}{The positive part of the expectation.}
\item{Eminus}{The negative part of the expectation.}
}
}
\description{
This function calculates the expectation of a distribution represented by a discrete CDF (Cumulative Distribution Function) of class "stepfun".
The expectation is computed along with the positive and negative contributions to the expectation.
}
\examples{
# Example usage with an ECDF
sf <- ecdf(mtcars$disp)
Ex(sf)
Extracting only the expected value
Ex(sf)$E

}
