% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATE.R
\name{PATE}
\alias{PATE}
\title{Partial Average Treatment Effect (PATE)}
\usage{
PATE(Y, Z, y.limit = NULL)
}
\arguments{
\item{Y}{A numeric vector of observed outcomes.}

\item{Z}{A numeric vector of treatment variables.}

\item{y.limit}{A numeric vector of length 2 specifying the limits for the outcome. Defaults to the minimum and maximum of \code{Y}.}
}
\value{
A list with the following elements:
\item{CTEemp}{The CTE function under ignorability assumption.}
\item{PATEemp}{The PATE function under ignorability assumption.}
\item{y}{A numeric vector of sequence values for the outcome range.}
\item{PATEempPlus}{The positive part of the PATE function under ignorability assumption.}
\item{PATEempMinus}{The negative part of the PATE function under ignorability assumption.}
}
\description{
This function estimates the Partial Average Treatment Effect (PATE), and the positive and negative parts of the PATE under ignorability assumption.
}
\examples{
# Example usage:
result <- PATE(Y = c(1, 2, 3, 4, 5), Z = c(0, 1, 0, 1, 1))

}
