% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.degreeoftruth.R
\name{plot.degreeoftruth}
\alias{plot.degreeoftruth}
\title{Calculate and plot probabilities of causal effect based on possible worlds}
\usage{
\method{plot}{degreeoftruth}(
  DegreesOfTruth,
  showplot = T,
  add = F,
  displayepsilon = T,
  addlegend = T,
  colors = c("blue3", "green3", "red3"),
  cex.legend = 1,
  legend.location = "topright",
  displaytitle = T,
  title.line = 1.2,
  lty = 1:length(DegreesOfTruth),
  lwd = 1,
  xlab = "Prior propensity p",
  ylab = "Probability",
  xmin = 0,
  xmax = 1,
  ymin = -0.02,
  ymax = 1.02,
  xlab.line = 2.5,
  ...
)
}
\arguments{
\item{add}{Logical value indicating whether to add to an existing plot (i.e., not creating a new plot). Default is \code{FALSE}.}

\item{displayepsilon}{Logical value indicating whether the \code{epsilon} threshold should be displayed in the plot. Default is \code{TRUE}.}

\item{addlegend}{Logical value indicating whether to add a legend to the plot. Default is \code{TRUE}.}

\item{colors}{A vector of colors used for plotting the three categories: No effect, Positive effect, and Negative effect. Default is \code{c("blue3", "green3", "red3")}.}

\item{cex.legend}{Size of legend. Default is \code{c(0, 1)}.}

\item{lty}{Line type for the plot. Default is \code{1}.}

\item{lwd}{Line width for the plot. Default is \code{1}.}

\item{ylab}{Label for the y-axis. Default is \code{"Probability"}.}

\item{\dots}{Additional graphical parameters passed to the plot function.}

\item{CTEpr}{A list containing the distribution of the CTE generated by the \code{CTEprob} function.}

\item{epsilon}{A threshold for considering the CTE as different from 0. Default is 0.}

\item{ylim}{y-axis limits. Default is \code{c(0, 1)}.}
}
\value{
A list containing the following probability vectors:
\itemize{
\item p_positive_effect: Probability of a positive causal effect.
\item p_negative_effect: Probability of a negative causal effect.
\item p_no_effect: Probability of no causal effect.
}
}
\description{
This function calculates and plots the probabilities of positive causal effect,
negative causal effect, and no causal effect based on prior success propensities for a
dichotomous outcome. The plot visualizes these probabilities across different prior
propensities, showing how the causal effect probabilities change with varying assumptions about
the likelihood of success.
}
\details{
The causal effect categories are defined over the space of all possible worlds as:
\itemize{
\item \eqn{E_{+, \epsilon} = \{ CTE > \epsilon \}} (Positive effect)
\item \eqn{E_{-, \epsilon} = \{ CTE < -\epsilon \}} (Negative effect)
\item \eqn{E_{0, \epsilon} = \{ |CTE| \leq \epsilon \}} (No effect)
}
The function calculates the probabilities of each of these events and plots them against the prior
propensity values.
}
