% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ecdfPI.R
\name{plot.ecdfPI}
\alias{plot.ecdfPI}
\title{Plot Method for ecdfPI Objects}
\usage{
\method{plot}{ecdfPI}(
  x,
  F1 = NULL,
  F2 = NULL,
  ...,
  ylab = "Fn(x)",
  verticals = FALSE,
  col.01line = "gray70",
  pch = 19,
  x.limit = NULL,
  col.bounds = "blue",
  alpha.bounds = 0.1
)
}
\arguments{
\item{x}{An object of class \code{ecdfPI}.}

\item{F1}{Lower bound of the ECDF. If \code{NULL}, bounds under no assumptions are used.}

\item{F2}{Upper bound of the ECDF. If \code{NULL}, bounds under no assumptions are used.}

\item{...}{Additional arguments passed to \code{plot.stepfun}.}

\item{ylab}{A title for the y-axis (default is \code{"Fn(x)"}).}

\item{verticals}{Logical; if \code{TRUE}, vertical lines are added at data points.}

\item{col.01line}{Color for the horizontal lines at 0 and 1 (default is \code{"gray70"}).}

\item{pch}{Plotting character or symbol to use (default is \code{19}).}

\item{x.limit}{Numeric vector of length 2 specifying the minimum and maximum x-values for plotting bounds.
Defaults to 20 times the range of the data if not specified.}

\item{col.bounds}{Color for the bands of identification (default is \code{"blue"}).}

\item{alpha.bounds}{Transparency level for the bands of partial identification (default is \code{0.1}).}
}
\description{
This function creates a plot for objects of class \code{ecdfPI}.
}
\details{
This function is a fork of \code{plot.ecdf} for \code{ecdfPI} objects.
}
