% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.quasiinverse.R
\name{plot.quasiinverse}
\alias{plot.quasiinverse}
\title{Plot Method for quasiinverse Objects}
\usage{
\method{plot}{quasiinverse}(
  x,
  F1 = NULL,
  F2 = NULL,
  ...,
  ylab = "Fn^{-1}(x)",
  verticals = FALSE,
  col.mmline = "gray70",
  pch = 19,
  x.limit = NULL,
  col.bounds = "blue",
  alpha.bounds = 0.1
)
}
\arguments{
\item{x}{An object of class \code{quasiinverse}.}

\item{...}{Additional arguments passed to \code{plot.stepfun}.}

\item{ylab}{A title for the y-axis (default is \code{"Fn(x)"}).}

\item{verticals}{Logical; if \code{TRUE}, vertical lines are added at data points.}

\item{col.mmline}{Color for the horizontal lines at minimum and maximum (default is \code{"gray70"}).}

\item{pch}{Plotting character or symbol to use (default is \code{19}).}
}
\description{
This function creates a plot for objects of class \code{quasiinverse}. It uses the base R function \code{plot.stepfun}
to plot the empirical cumulative distribution function (ECDF).
}
\details{
This function is a fork of \code{plot.ecdf} for \code{ecdfPI} objects.
}
