% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCTEprob.R
\name{plotCTEprob}
\alias{plotCTEprob}
\title{Calculate and plot probabilities of causal effect based on possible worlds}
\usage{
plotCTEprob(
  CTEpr,
  alpha = 0,
  displayalpha = T,
  addlegend = T,
  add = F,
  colors = c("blue3", "green3", "red3"),
  lty = 1,
  lwd = 1,
  ylab = "Probability",
  ylim = c(0, 1),
  ...
)
}
\arguments{
\item{CTEpr}{A list containing the distribution of the CTE generated by the \code{CTEprob} function.}

\item{alpha}{A threshold for considering the CTE as different from 0. Default is 0.}

\item{displayalpha}{Logical value indicating whether the \code{alpha} threshold should be displayed in the plot. Default is \code{TRUE}.}

\item{addlegend}{Logical value indicating whether to add a legend to the plot. Default is \code{TRUE}.}

\item{add}{Logical value indicating whether to add to an existing plot (i.e., not creating a new plot). Default is \code{FALSE}.}

\item{colors}{A vector of colors used for plotting the three categories: No effect, Positive effect, and Negative effect. Default is \code{c("blue3", "green3", "red3")}.}

\item{lty}{Line type for the plot. Default is \code{1}.}

\item{lwd}{Line width for the plot. Default is \code{1}.}

\item{ylab}{Label for the y-axis. Default is \code{"Probability"}.}

\item{ylim}{y-axis limits. Default is \code{c(0, 1)}.}

\item{\dots}{Additional graphical parameters passed to the plot function.}
}
\value{
A list containing the following probability vectors:
\itemize{
\item p_positive_effect: Probability of a positive causal effect.
\item p_negative_effect: Probability of a negative causal effect.
\item p_no_effect: Probability of no causal effect.
}
}
\description{
This function calculates and plots the probabilities of positive causal effect,
negative causal effect, and no causal effect based on prior success propensities for a
dichotomous outcome. The plot visualizes these probabilities across different prior
propensities, showing how the causal effect probabilities change with varying assumptions about
the likelihood of success.
}
\details{
The causal effect categories are defined over the space of all possible worlds as:
\itemize{
\item \eqn{E_{+, \alpha} = \{ CTE > \alpha \}} (Positive effect)
\item \eqn{E_{-, \alpha} = \{ CTE < -\alpha \}} (Negative effect)
\item \eqn{E_{0, \alpha} = \{ |CTE| \leq \alpha \}} (No effect)
}
The function calculates the probabilities of each of these events and plots them against the prior
propensity values.
}
