% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/possibleworlds.R
\name{possibleworlds}
\alias{possibleworlds}
\title{Calculate the distribution of the Causal Treatment Effect (CTE) for a Dichotomous Outcome}
\usage{
possibleworlds(
  M,
  nNAZ1 = 0,
  nNAZ0 = 0,
  nNA = NULL,
  nZ1K1,
  nZ0K1,
  nY1Z1K1,
  nY1Z0K1,
  p,
  dec.prec = 13,
  alpha = 0.05,
  approxtonorm = FALSE
)
}
\arguments{
\item{M}{Total number of individuals in the population.}

\item{nNAZ1, nNAZ0}{Use only if nNA is NULL. Number of missing data in the observed outcome for Zobs=1 and Zobs=0.}

\item{nNA}{Number of missing data in the observed outcome. If nNA is NULL, nNAZ1+nNAZ0 is used.}

\item{nZ1K1}{Number of statistical units with Z=1 and K=1.}

\item{nZ0K1}{Number of statistical units with Z=0 and K=1.}

\item{nY1Z1K1}{Number of statistical units with Y=1, Z=1, and K=1.}

\item{nY1Z0K1}{Number of statistical units with Y=1, Z=0, and K=1.}

\item{p}{Propensity values for which the CTE distribution is calculated.}

\item{dec.prec}{Number of decimal places for precision in the calculations (default is 13).}

\item{alpha}{Significance of the credible interval of Pi.}

\item{approxtonorm}{If approxtonorm is TRUE, normal approximation of the distribution of Pi is used.}
}
\value{
A list containing the CTE(0) distribution for a dichotomous outcome in the measurable space. The list includes:

\item{masses}{A data frame with the probability mass \eqn{P(CTE(0) = cte)}.}
\item{cdf}{A data frame with the cumulative probability \eqn{P(CTE(0) \leq cte)}.}
\item{probMZ}{Probabilities identified from the data:}
\itemize{
\item{probK1Z1}{Probability \eqn{P(K = 1 | Z = 1)}.}
\item{probK1Z0}{Probability \eqn{P(K = 1 | Z = 0)}.}
\item{probY1Z1K1}{Probability \eqn{P(Y = 1 | Z = 1, K = 1)}.}
\item{probY1Z0K1}{Probability \eqn{P(Y = 1 | Z = 0, K = 1)}.}
\item{p0}{The prior propensity where the probability of a positive causal effect equals the probability of a negative causal effect.}
}
\item{probassing}{A vector containing probabilies of assignment to treatments Z=1 and Z=0.}Evangelion 3.0+1.0 (2021)
\item{Pi}{Probabilities identified from the data:}
\itemize{
\item{EpPi}{Expected value of \eqn{\Pi}.}
\item{CDF_Pi}{CDF function of \eqn{\Pi}, \eqn{F_{\Pi,p}}.}
\item{qPileft}{Left-continuous quasi-inverse of \eqn{F_{\Pi,p}}.}
\item{qPiright}{Right-continuous quasi-inverse of \eqn{F_{\Pi,p}}.}
\item{q0025left}{Quantile 0.025 of \eqn{\Pi} using the left-continuous quasi-inverse.}
\item{q0025right}{Quantile 0.025 of \eqn{\Pi} using the right-continuous quasi-inverse.}
\item{q0975left}{Quantile 0.975 of \eqn{\Pi} using the left-continuous quasi-inverse.}
\item{q0975right}{Quantile 0.975 of \eqn{\Pi} using the right-continuous quasi-inverse.}
\item{q0025}{Lower bound of the 95\\% credible interval of \eqn{Pi}.}
\item{q0975}{Upper bound of the 95\\% credible interval of \eqn{Pi}.}
}
}
\description{
This function calculates the CTE(0) distribution for a dichotomous outcome over all possible worlds
that are consistent with the observed data.
}
\details{
The variable \eqn{\mathcal{M}} represents the set of labels for the individuals in the population of interest.
Observed causal inference data are typically presented as variables defined for each individual \eqn{m \in M}.
The outcome \eqn{Y} is defined over the space \eqn{\mathcal{M} \times \{0, 1\}}, where \eqn{K(m, z) = 1} if \eqn{Y(m, z)} is observed,
and 0 otherwise. \eqn{Z} is the projection of \eqn{(m, z)} in \eqn{Z}, meaning \eqn{Z(m, z) = z}.

The outcome \eqn{Y} is dichotomous, and each missing outcome value can take either 0 or 1.
Consider the space of all possible worlds where for each world, there exists an outcome \eqn{Y^*},
where \eqn{Y^*} coincides with the observed outcome \eqn{Y_{\text{obs}}} for the observed statistical units.
For each \eqn{Y^*}, the CTE(0) is computed. The CTE(0) is a random variable over all possible worlds.

The value \eqn{p} represents the prior probability that each unobserved outcome is 1. This prior probability is denoted as
the propensity for success. Assuming a common \eqn{p} for all unobserved outcomes implies the absence of a causal effect.
Based on \eqn{p}, the probability of each possible world is determined. If all possible worlds are equally likely, \eqn{p = 0.5},
which assumes no causal effect. This assumption becomes unrealistic when the outcome's success probability is significantly
different from 0.5.
}
\examples{
# Example usage of the CTEprob function
result <- CTEprob(M = 200, nNA = 0, nZ1K1 = 65, nZ0K1 = 135, 
                  nY1Z1K1 = 34, nY1Z0K1 = 91, p = c(0:100)/100 )
result$masses
result$cdf

}
