% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quasiinverse.R
\name{quasiinverse}
\alias{quasiinverse}
\title{Calculate quasi-inverse of discrete cumulative distribution functions}
\usage{
quasiinverse(f, continuity = "right", ymin = NULL, ymax = NULL)
}
\arguments{
\item{f}{A function of class \code{stepfun}, \code{ecdf} or \code{ecdfPI}.}

\item{continuity}{Continuity side (default: right).}

\item{ymin, ymax}{Minimum and maximum values of the random variable. If \code{NULL}, the data’s minimum and maximum values are used (\code{NULL} by default).}
}
\value{
A function of class \code{quasiinverse}.
}
\description{
This function calculates the quasi-inverse of discrete cumulative distribution function.
}
\examples{
result <- ecdfPI(iris$Sepal.Length)
plot( quasiinverse(result, 'left') , col="blue" , 
      main = "Quasi-inverses of iris$Sepal.Length" ,
      xlab = "q" , ylab = "y")
lines( quasiinverse(result, 'right') , col="brown")

legend("topleft", 
       legend = c(expression(F^{-1} * (q) == "inf { y : F(y) >= q}"),
                  expression(F^{(-1)} * (q) == "sup { y : F(y) <= q}")),
       col = c("brown", "blue"), 
       lwd = 2, 
       box.lty = 0) 

}
