% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportU.R
\name{supportU}
\alias{supportU}
\title{Compute Support for \eqn{U = Y_1 - Y_0}}
\usage{
supportU(Y, Z, y1.limit = NULL, y0.limit = NULL)
}
\arguments{
\item{Y}{A numeric vector of observed outcomes.}

\item{Z}{A numeric vector indicating the treatment variable (e.g., 1 for treatment and 0 for control).}

\item{y1.limit}{A vector of length 2 specifying the minimum and maximum values of Y1. If NULL,
it defaults to the range of Y in the treatment group.}

\item{y0.limit}{A vector of length 2 specifying the minimum and maximum values of Y0. If NULL,
it defaults to the range of Y in the control group.}
}
\value{
A list containing:
\item{ suppY1 }{ Sorted values of Y for the treatment group. }
\item{ suppY0 }{ Sorted values of Y for the control group. }
\item{ suppU }{ Sorted values of U = Y1 - Y0. }
\item{ suppXYU }{ A data frame with all combinations of Y1 and Y0, and their differences (U). }
}
\description{
This function computes the support for the outcome difference \eqn{U = Y_1 - Y_0}, where \eqn{Y_1} and \eqn{Y_0} are outcomes
corresponding to treatment and control groups, respectively.
}
