% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WilliamsonDowns1990.R
\name{WilliamsonDowns1990}
\alias{WilliamsonDowns1990}
\title{Williamson and Downs (1990) bounds for the distribution of the difference of counterfactual outcomes}
\usage{
WilliamsonDowns1990(
  Y,
  Z,
  y1.limit = NULL,
  y0.limit = NULL,
  boundaries = "4epsilon4eta",
  gridsize = 2000,
  delta.u = diff(range(Y[Z == 1], na.rm = T) - rev(range(Y[Z == 0], na.rm = T)))/gridsize,
  delta.ldbudboptim = diff(range(Y[Z == 1], na.rm = T) - rev(range(Y[Z == 0], na.rm =
    T)))/gridsize,
  includesupport = FALSE,
  addtoU = NULL,
  parallel = FALSE,
  freeCores = 1,
  eps11 = 0,
  eps12 = 0,
  eps01 = 0,
  eps02 = 0,
  eta11 = 0,
  eta12 = 0,
  eta01 = 0,
  eta02 = 0
)
}
\arguments{
\item{Y}{A numeric vector of observed outcomes.}

\item{Z}{A dummy numeric vector indicating the treatment assignment (1 for treatment, 0 for control).}

\item{y1.limit, y0.limit}{Numeric vectors of length 2, specifying the lower and upper bounds for the potential outcomes under treatment 1 and 0. If \code{NULL}, the bounds are automatically set based on the observed treatment outcomes.}

\item{boundaries}{A character vector specifying the boundary methods to use. Options include "Ignorability", "NoAssumptions", and "FourEpsilon".}

\item{gridsize}{Number of elements in the sequences for bounds evaluation (values of \eqn{U = Y_{1} - Y_{0}}) and the grid for minimum-convolution. Default is 2000. \code{gridsize} is ignored for the respective grid if \code{delta.u} or \code{delta.ldbudboptim} is specified.}

\item{delta.u}{A scalar defining the spacing of the sequence for \eqn{U = Y_{1} - Y_{0}} values. By default, \code{delta.u} is determined using \code{gridsize} and the empirical range of \eqn{U = Y_{1} - Y_{0}}.}

\item{delta.ldbudboptim}{A scalar defining the spacing for the sequence used in the minimum-convolution steps. By default, \code{delta.ldbudboptim} is determined using \code{gridsize} and the empirical range of \eqn{U = Y_{1} - Y_{0}}.}

\item{includesupport}{Logical value indicating whether to include the set \eqn{\{y_1 - y_0 : y_1 \in Y[Z == 1] \wedge y_0 \in Y[Z == 0]\}} in the sequence of values for \eqn{U = Y_{1} - Y_{0}} constructed with \code{delta.u}. Default is \code{FALSE}.}

\item{addtoU}{Optional numeric vector containing specific values to be added to the grid for evaluating the bounds. Default is \code{NULL}.}

\item{parallel}{A logical indicating whether parallel computation should be used (if \code{TRUE}, the \code{parallel} library will be employed). Default is \code{FALSE}.}

\item{freeCores}{The number of CPU cores to reserve for other tasks when \code{parallel = TRUE}. Default is 1.}

\item{eps11, eps12, eps01, eps02}{Scalars defining the epsilon parameters for the "FourEpsilon" boundary method. Default values are 0.}

\item{eta11, eta12, eta01, eta02}{Scalars defining the eta parameters for the 4-\eqn{\eta} assumption. Default values are 0.}
}
\description{
Implementation of the bounds for \eqn{U = Y_{1} - Y_{0}} from Williamson and Downs (1990) using partial identification bounds for the counterfactual distributions. The function returns step functions for the lower and upper bounds of the treatment effect distribution, \eqn{P(Y_{1} - Y_{0} \leq y)}, under different assumptions.
}
\details{
The bounds are computed over a grid of values determined by the \code{delta} parameter. When \code{includesupport} is TRUE, the grid also incorporates all possible values of \eqn{U} based on the observed data. The \code{delta.ldbudboptim} parameter specifies the grid on which the minimum-convolution is evaluated.

One way to relax ignorability assumptions is by introducing tolerance for discrepancies between identified and non-identified distributions. The \code{FourEpsilon} boundaries are based on the following identification assumptions: .

\deqn{
-\varepsilon_{11} 
\leq 
P(Y_{z_1} \leq y | Z=z_1) - P(Y_{z_1} \leq y | Z=z_0) 
\leq 
\varepsilon_{12}
\,,
}
\deqn{
-\varepsilon_{01} 
\leq 
P(Y_{z_0} \leq y | Z=z_1) - P(Y_{z_0} \leq y | Z=z_0) 
\leq 
\varepsilon_{02}
\,,
}

where \eqn{\varepsilon_{11}, \varepsilon_{12}, \varepsilon_{01}, \varepsilon_{02}} correspond to \code{eps11}, \code{eps12}, \code{eps01}, and \code{eps02},, respectively.

Warning: The minimum-convolution strongly depends on the delta and delta.ldbudboptim parameters. If delta and delta.ldbudboptim are too large, the estimation will be poor. Be cautious: as delta (or delta.ldbudboptim) decreases, the computation time increases significantly.
}
