% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctefunctions.R
\name{ctefunctions}
\alias{ctefunctions}
\title{Partial Identification bounds for counterfactual distributions}
\usage{
ctefunctions(
  Y,
  Z,
  y1.limit = NULL,
  y0.limit = NULL,
  eps11 = 1,
  eps12 = 1,
  eps01 = 1,
  eps02 = 1,
  eta11 = 1,
  eta12 = 1,
  eta01 = 1,
  eta02 = 1,
  na.rm = F
)
}
\arguments{
\item{Y}{A numeric vector of observed outcomes.}

\item{Z}{A binary treatment indicator (0 or 1).}

\item{y1.limit}{A numeric vector of length 2 specifying the lower and upper bounds of the
support for \eqn{Y_1}. Default is NULL, in which case the bounds are
based on the observed data.}

\item{y0.limit}{A numeric vector of length 2 specifying the lower and upper bounds of the
support for \eqn{Y_0}. Default is NULL, in which case the bounds are
based on the observed data.}

\item{eps11, eps12, eps01, eps02}{Numeric values specifying the bound for 4-epsilon assumption. Default is 1.}

\item{eta11, eta12, eta01, eta02}{Numeric values specifying the bound for 4-eta assumption. Default is 1.}

\item{na.rm}{A logical value indicating whether to remove missing values from the dataset.
Default is FALSE.}
}
\value{
A list containing:
\itemize{
\item \code{ecdfYZ1} A step function representing the empirical cumulative distribution
function of \eqn{Y_1} given \eqn{Z=1}.
\item \code{ecdfYZ0} A step function representing the empirical cumulative distribution
function of \eqn{Y_0} given \eqn{Z=0}.
\item \code{pZ1} The probability \eqn{P(Z = 1)}.
\item \code{pZ0} The probability \eqn{P(Z = 0)}.
\item \code{CTE} The conditional treatment effect (CTE) under ignorability assumption.
\item \code{CTEplus} The positive part of the conditional treatment effect (CTE) under ignorability assumption.
\item \code{CTEminus} The negative part of the conditional treatment effect (CTE) under ignorability assumption.
\item \code{yz1} The sorted values of \eqn{Y} for units with \eqn{Z = 1}.
\item \code{yz0} The sorted values of \eqn{Y} for units with \eqn{Z = 0}.
\item \code{ecdfPIbounds} A list of partial identification bounds for the counterfactual distributions
with no assumptions.
\item \code{ecdfPIbounds.eps} A list of partial identification bounds for the counterfactual distributions
under 4-epsilon and 4-eta assumptions.
\item \code{dataset} The dataset used for the analysis.
}
}
\description{
This function computes partial identification bounds for the counterfactual outcomes,
both with no assumptions, under ignorability and under the 4-epsilon assumption.
It returns step functions bounds and the CTE under ignorability assumption.
}
\details{
One way to relax ignorability assumptions is by introducing tolerance for discrepancies between identified and non-identified distributions. The - \strong{FourEpsilon} boundaries are based on the following identification assumptions: .

\deqn{
-\varepsilon_{11} 
\leq 
P(Y_{1} \leq y | Z=1) - P(Y_{1} \leq y | Z=0) 
\leq 
\varepsilon_{12}
\,,
}
\deqn{
-\varepsilon_{01} 
\leq 
P(Y_{0} \leq y | Z=1) - P(Y_{0} \leq y | Z=0) 
\leq 
\varepsilon_{02}
\,,
}

where \eqn{\varepsilon_{11}, \varepsilon_{12}, \varepsilon_{01}, \varepsilon_{02}} correspond to \code{eps11}, \code{eps12}, \code{eps01}, and \code{eps02}, respectively.

If the observed outcome has missing data, assumption 4-eta is considered:

\deqn{
-\eta_{11} 
\leq 
P(Y_{1} \leq y | Z=1, W=1) - P(Y_{1} \leq y | Z=1, W=0) 
\leq 
\eta_{12}
}

\deqn{
-\eta_{01} 
\leq 
P(Y_{0} \leq y | Z=0, W=1) - P(Y_{0} \leq y | Z=0, W=0) 
\leq 
\eta_{02}
}

where \eqn{\eta_{11}, \eta_{12}, \eta_{01}, \eta_{02}} correspond to \code{eta11}, \code{eta12}, \code{eta01}, and \code{eta02}, respectively.
}
\examples{
# Example usage
set.seed(123)
Y <- rnorm(100)
Z <- rbinom(100, 1, 0.5)
result <- ctefunctions(Y, Z, y1.limit=c(-1, 2), y0.limit=c(-2, 3))


}
