% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate.sf.R
\name{integrate.sf}
\alias{integrate.sf}
\title{Cumulative integral for step functions}
\usage{
integrate.sf(sf)
}
\arguments{
\item{sf}{A step function of class "stepfun".}
}
\value{
A function of class "sfintegrate" and "stepfun", which represents the cumulative integral of the input step function.
}
\description{
This function returns the cumulative integral of a step function.
}
\examples{
# Example of a step function
x <- c(0, 1, 2, 3, 4)
y <- c(0, 1, 0, 1, 0)
sf <- stepfun(x, y)

# Integrate the step function
integral_sf <- integrate.sf(sf)

# Plot the original step function and its integral
plot(sf, main = "Step Function")
plot(integral_sf, main = "Integral of Step Function")


}
